requirements.txt
Flask==3.0.3
gunicorn==23.0.0
google-cloud-firestore==2.21.0
firebase-admin==6.7.0